DAC-1974 Split application into multiple deployments (#413)
Split IaC code into two applications with their own subdirectories
Update GitHub workflows to deploy and teardown the separate applications
Update scripts, husky hooks and npm scripts to handle IaC split
Update README to cover modified repositoru structure and deployments
Update various GitHub action paths and versions
Update todos and development-decisions

+++ scripts/build-sam-template.sh
@@ -1,14 +1,24 @@
 #!/bin/bash -e
 
+APP_NAME=$1
+
+if [ -z "$APP_NAME" ] || { [ "$APP_NAME" != "main" ] && [ "$APP_NAME" != "quicksight-access" ]; }; then
+  echo "App name must be provided and be one of 'main' or 'quicksight-access'"
+  echo "Usage (bash): build-sam-templates.sh APP_NAME"
+  echo "Usage (npm): npm run iac:build -- APP_NAME"
+  exit 1
+fi
+
 ROOT=$(pwd)
 IAC_DIR="$ROOT/iac"
-RESOURCE_DIR="$IAC_DIR/resources"
+BASE_FILE="$IAC_DIR/$APP_NAME/base.yml"
+RESOURCE_DIR="$IAC_DIR/$APP_NAME/resources"
 OUTFILE="$ROOT/template.yaml"
 
 rm -f "$OUTFILE"
 
 # start the template with the initial section in base.yml, and begin the Resources section
-cat "$IAC_DIR"/base.yml > "$OUTFILE"
+cat "$BASE_FILE" > "$OUTFILE"
 echo -e "\nResources:" >> "$OUTFILE"
 
 for file in "$RESOURCE_DIR"/*; do
@@ -16,4 +26,4 @@ for file in "$RESOURCE_DIR"/*; do
   echo >> "$OUTFILE"
 done
 
-prettier "$OUTFILE" --write
+prettier "$OUTFILE" --write > /dev/null
